move comps
