+++ src/components/Admin/Teams.jsx
@@ -6,7 +6,7 @@ import Team from "./Team";
 import Filters from "@/components/Admin/Filters";
 import SortIcon from "./SortIcon";
 import Toolbar from "@/components/Admin/Toolbar";
-import Title from "../Title";
+import Title from "./Title";
 const teams = [
   {
     uid: 1,
