--- src/components/Form.jsx
@@ -1,5 +1,0 @@
-const Form = () => {
-  return <div>Form</div>;
-};
-
-export default Form;
