+++ src/components/Forms/Admin.jsx
@@ -6,7 +6,7 @@ import Radio from "@/components/Radio";
 import Checkbox from "../Checkbox";
 import Input from "../Input";
 import { Row, Col } from "react-bootstrap";
-import Button from "../Button";
+import Button from "./Button";
 
 const affiliations = [
   "ops",
